Render proper homepage component based on landing page mode
