+++ pages/index.js
@@ -1,84 +1,20 @@
-import React, { useEffect, useState } from 'react';
-import dynamic from 'next/dynamic';
 import { hasuraStreamArticles } from '../lib/homepage.js';
 import { cachedContents } from '../lib/cached';
 import { hasuraGetHomepageEditor } from '../lib/articles.js';
 import { getArticleAds } from '../lib/ads.js';
 import { hasuraLocaliseText } from '../lib/utils.js';
-import Layout from '../components/Layout';
-import ArticleStream from '../components/homepage/ArticleStream';
-import Placeholder from '../components/homepage/Placeholder';
-import homepageStyles from '../styles/homepage.js';
 
-const BigFeaturedStory = dynamic(() =>
-  import(`../components/homepage/BigFeaturedStory`)
-);
-const LargePackageStoryLead = dynamic(() =>
-  import(`../components/homepage/LargePackageStoryLead`)
-);
+import Homepage from '../components/Homepage';
+import LandingPage from '../components/LandingPage';
 
-export default function Home({
-  selectedLayout,
-  featured,
-  topFeatured,
-  bottomFeatured,
-  streamArticles,
-  sections,
-  locale,
-  siteMetadata,
-  expandedAds,
-}) {
-  const [featuredArticle, setFeaturedArticle] = useState(featured);
-  const [subFeaturedTopArticle, setSubFeaturedTopArticle] = useState(
-    topFeatured
+export default function Home(props) {
+  const component = props.siteMetadata.landingPage ? (
+    <LandingPage {...props} />
+  ) : (
+    <Homepage {...props} />
   );
-  const [subFeaturedBottomArticle, setSubFeaturedBottomArticle] = useState(
-    bottomFeatured
-  );
-  const [mostRecentArticles, setMostRecentArticles] = useState(streamArticles);
 
-  return (
-    <div className="homepage">
-      <Layout meta={siteMetadata} sections={sections} locale={locale}>
-        {!selectedLayout && <Placeholder />}
-        {selectedLayout.name === 'Big Featured Story' && (
-          <BigFeaturedStory
-            sections={sections}
-            featuredArticle={featuredArticle}
-            setFeaturedArticle={setFeaturedArticle}
-            isAmp={false}
-            metadata={siteMetadata}
-          />
-        )}
-        {selectedLayout.name === 'Large Package Story Lead' && (
-          <LargePackageStoryLead
-            locale={locale}
-            featuredArticle={featuredArticle}
-            setFeaturedArticle={setFeaturedArticle}
-            subFeaturedTopArticle={subFeaturedTopArticle}
-            setSubFeaturedTopArticle={setSubFeaturedTopArticle}
-            subFeaturedBottomArticle={subFeaturedBottomArticle}
-            setSubFeaturedBottomArticle={setSubFeaturedBottomArticle}
-            sections={sections}
-            isAmp={false}
-            metadata={siteMetadata}
-          />
-        )}
-        <ArticleStream
-          articles={mostRecentArticles}
-          sections={sections}
-          showCategory={true}
-          isAmp={false}
-          title={siteMetadata.homepageArticleStreamHed}
-          metadata={siteMetadata}
-          ads={expandedAds}
-        />
-      </Layout>
-      <style jsx global>
-        {homepageStyles}
-      </style>
-    </div>
-  );
+  return component;
 }
 
 export async function getStaticProps({ locale }) {
